Add CurrentUser abstraction and Business logic related to the Role

+++ Application/Commands/DeleteEmployeeCommand.cs
@@ -17,10 +17,12 @@ namespace Application.Commands
     public class DeleteEmployeeCommandHandler : IRequestHandler<DeleteEmployeeCommand>
     {
         private readonly IEmployeeRepository _repository;
+        private readonly ICurrentUser _currentUser;
 
-        public DeleteEmployeeCommandHandler(IEmployeeRepository repository)
+        public DeleteEmployeeCommandHandler(IEmployeeRepository repository, ICurrentUser currentUser)
         {
             _repository = repository;
+            _currentUser = currentUser;
         }
 
         public async Task Handle(DeleteEmployeeCommand request, CancellationToken cancellationToken)
@@ -30,6 +32,11 @@ namespace Application.Commands
 
             if (request.Id == Guid.Empty)
                 throw new ValidationException("Invalid Employee Id.");
+
+            var data = await _repository.GetByIdAsync(request.Id, cancellationToken);   
+
+            if (data?.Role >= _currentUser.Role)
+                throw new BusinessRuleException("Unauthorized to remove Employee with higher role level.");
 
             await _repository.DeleteAsync(request.Id, cancellationToken);
             return;
